+++ core/src/test/java/arez/ArezTest.java
@@ -21,7 +21,6 @@ public class ArezTest
   public void zone_basicOperation()
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
 
     assertEquals( ArezZoneHolder.getDefaultZone().getContext(), Arez.context() );
     assertEquals( ArezZoneHolder.getZoneStack().size(), 0 );
@@ -46,7 +45,6 @@ public class ArezTest
   public void zone_multipleZones()
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
 
     assertEquals( ArezZoneHolder.getDefaultZone().getContext(), Arez.context() );
     assertEquals( ArezZoneHolder.getZoneStack().size(), 0 );
@@ -178,7 +176,6 @@ public class ArezTest
   public void deactivateZone_whenNotActive()
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
     assertInvariantFailure( () -> Arez.deactivateZone( new Zone() ),
                             "Arez-0004: Attempted to deactivate zone that is not active." );
   }