+++ core/src/test/java/arez/ZoneTest.java
@@ -12,7 +12,6 @@ public class ZoneTest
   public void zone_safeRun_Function()
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
 
     final Zone zone1 = Arez.createZone();
 
@@ -38,7 +37,6 @@ public class ZoneTest
   public void zone_safeRun_Procedure()
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
 
     final Zone zone1 = Arez.createZone();
 
@@ -60,7 +58,6 @@ public class ZoneTest
   public void zone_run_Function_throwsException()
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
 
     final Zone zone1 = Arez.createZone();
 
@@ -83,7 +80,6 @@ public class ZoneTest
   public void zone_run_Procedure_throwsException()
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
 
     final Zone zone1 = Arez.createZone();
 
@@ -108,7 +104,6 @@ public class ZoneTest
     throws Throwable
   {
     ArezTestUtil.enableZones();
-    ArezTestUtil.resetState();
 
     final Zone zone1 = Arez.createZone();